requirements.txt

djangorestframework

django-filter

drf-spectacular

djangorestframework-simplejwt

# XML support
 djangorestframework-xml
